promhttp: handle a Mapping for a metric family name that doesn't conform to the openmetrics naming
If you previously had a metric we were transforming that ended in `_count`,
`_bucket`, or `_sum`, we assumed these were a part of a histogram
and stripped this suffix to nornmalize into the correct name.

According to the spec, naming a normal counter or whatever that ends in
`_count` is strongly discouraged and is considered wrong.

But to accommodate, it's not our place to enforce that naming here.
Instead, if there's an explicit mapping to the exact match, we favor
that first before normalizing.